% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjMatrix.R
\name{adjMatrix}
\alias{adjMatrix}
\title{A Function for calculate an adjacency matrix from a dataframe}
\usage{
adjMatrix(dataframe, boolean = TRUE)
}
\arguments{
\item{a}{three columns dataframe and a logical output selector}
}
\description{
This function takes the first two columns of a distance dataframe or matrix and return it's adjacency matrix. You can also select a logical or numerical output.
}
\examples{
> initialVertex <- c("A","A","C","C")
> finalVertex <- c("B","C","D","E")
> vertexDistance <- c(2,3,1,4)
> data <- data.frame(initialVertex, finalVertex, vertexDistance)
> data
initialVertex finalVertex vertexDistance
1             A           B              2
2             A           C              3
3             C           D              1
4             C           E              4
> matriz <- adjMatrix(data)
> matriz
        A     B     C     D     E
[1,]  TRUE  TRUE FALSE FALSE FALSE
[2,]  TRUE FALSE  TRUE FALSE FALSE
[3,] FALSE FALSE  TRUE  TRUE FALSE
[4,] FALSE FALSE  TRUE FALSE  TRUE

> matriz <- adjMatrix(datos, F)
> matriz
    A B C D E
[1,] 1 1 0 0 0
[2,] 1 0 1 0 0
[3,] 0 0 1 1 0
[4,] 0 0 1 0 1
}
\keyword{adjacency}
\keyword{matrix}
